refactor(ragChatModal): move RAGChatModal to src directory
- Relocate ragChatModal.ts to src directory
- Update import paths to reflect new location
- Maintain existing functionality without changes
- Ensure consistent formatting and styling
